% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data Set}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{The distance the fire covered (numeric)}
  \item{DAMAGE}{The damage caused by the wildfire (numeric)}
}
}
\source{
\url{https://canvas.ou.edu/courses/409964/assignments/3067006}
}
\usage{
fire
}
\description{
This data set contains data relating to the distance the wildfire traveled and the damage it caused.
}
\keyword{datasets}
