% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_maxlikg.R
\name{my_maxlikg}
\alias{my_maxlikg}
\title{my_maxlikg}
\usage{
my_maxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{a function selected, usually a logarithmic function}

\item{theta}{an input of the mean/certain value}
}
\value{
a plot of the maximum likelihood of a joint binomial and Poisson density.
}
\description{
my_maxlikg
}
\examples{
\dontrun{mymaxlikg(theta= seq(1,10, length= 1000))}
}
