% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynotch.R
\name{my_notchplot}
\alias{my_notchplot}
\title{Notched Boxplot}
\usage{
my_notchplot(data, var, xlab = NULL, main = NULL, col = "magenta")
}
\arguments{
\item{data}{a data table}

\item{var}{the variable in question}

\item{xlab}{x label, defaults to column name}

\item{main}{title of the plot}

\item{col}{purple}
}
\value{
a notched boxplot
}
\description{
Notched Boxplot
}
\examples{
\dontrun{my_notchplot(spruce, BHDiameter)}
}
