% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = mean, alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{the sample}

\item{fun}{the function to be used}

\item{alpha}{alpha based on the percent interval}

\item{cx}{text size for plots}

\item{...}{any additional parameters}
}
\value{
a histogram with the bootstrap interval on it
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter= 10000, x= sam, fun= mean, alpha= 0.05)}
}
