% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb1}
\alias{mycltb1}
\title{Title}
\usage{
mycltb1(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{the sample size}

\item{iter}{the number of iterations}

\item{p}{the probability}

\item{...}{any other parameters}
}
\value{
a histogram of a binomial distribution with the density and theoretical
}
\description{
Title
}
\examples{
\dontrun mycltb1(n= 10, iter= 10000, p= 0.7)
}
