% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My First Function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A list of components x and y with y being the square of x
}
\description{
My First Function
}
\examples{
myfirstfunction(1:10)
}
