% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{a vector}

\item{sigma}{a vector}

\item{a}{av vector}
}
\value{
a list and plot given mu, sigma, and a and generates probabilites off of a
}
\description{
Title
}
