% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, p, gamma)
}
\arguments{
\item{N}{the number of available seats on the plane}

\item{p}{the probability that a passenger will show}

\item{gamma}{the probability that the flight will be overbooked}
}
\value{
The plots of the descrete and continuous distributions, list of values
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N= 400, p = 0.95, gamma = 0.02)}
}
