% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marbles.R
\name{sample_marbles}
\alias{sample_marbles}
\title{Marble sampling iterations}
\usage{
sample_marbles(iter = 1, size = 5, white = 12, black = 8)
}
\arguments{
\item{iter}{#iterations}

\item{size}{#n}

\item{white}{#12 marbles}

\item{black}{#8 marbles}
}
\value{
#the iterations of that sample
}
\description{
Marble sampling iterations
}
\examples{
#sample_marbles(iter=4)
}
